Rework suspend state machine to support asynchronous suspend
Right now it's the only way we support suspending the system but when required the rework
to suspend both ways (synchronous and asynchronous) through nyx should be rather small.

Requires an extension of the official nyx API to support the two new methods needed
to suspend asynchronously and resume afterwards.

Signed-off-by: Simon Busch <morphis@gravedo.de>
Signed-off-by: Herman van Hazendonk <github.com@herrie.org>