Don't block main thread when in sleep state
When working with synchronous suspend/resume it's fine to block the other thread from
processing but not when we're working asynchronous. Especially when we're in a low power
state and the activity manager requests a activity to be started we currently deny that
instead of of blocking the main thread until the device has left the low power state and
is back to normal operations. This can be improved later to allow certain activities also
in low power mode to get a better battery lifetime.

Signed-off-by: Simon Busch <morphis@gravedo.de>
Signed-off-by: Herman van Hazendonk <github.com@herrie.org>